requirements.txt
# Core
python-dotenv
pyyaml
pytest
requests
openai

# Optional for later visualization or UI
gradio
